Update game over screen to use GenericMenuEvents
