--- Pop Pixie/Assets/Core Scripts/Game Over.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 001a614cd7b3d4b9da403f8266eb6d04
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
