+++ Pop Pixie/Assets/Other Scenes/Game Over.unity
@@ -576,6 +576,11 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   Fader: {fileID: 739071572}
   Music: {fileID: 8300000, guid: 803478d837d6e40079d7f3361787f199, type: 3}
+  AudioId: game over
+  FadeInDelay: 0.5
+  FadeInDuration: 1
+  FadeOutDuration: 0.5
+  PostFadeOutDelay: 0
 --- !u!224 &1086275790
 RectTransform:
   m_ObjectHideFlags: 0
